backend_repo/backend/requirements.txt
Django
django-cors-headers
djangorestframework
drf-yasg
requests
environs
gunicorn

